refactor input related classes
